pyproject.toml
[tool.poetry]
name = "storytime_ai"
version = "0.3.0"
description = "View and create interactive stories in terminal"
authors = ["Dronakurl <konrad2710@gmail.com>"]
readme = "README.rst"
license = "GPL-3.0-or-later"
keywords = ["interactive", "terminal", "textual", "game", "openai"]
homepage = "https://github.com/Dronakurl/storytime/"
repository = "https://github.com/Dronakurl/storytime/"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: GNU General Public License (GPL)",
    "Operating System :: OS Independent",
    "Topic :: Games/Entertainment :: Role-Playing",
]


[tool.poetry.dependencies]
python = "^3.11"
colorlog = "^6.7.0"
jinja2 = "^3.1.2"
pandas = "^2.0.3"
python-dotenv = "^1.0.0"
streamlit = {version = "^1.26.0", optional = true}
networkx = {version="^3.1", optional = true }
openai = {version="^0.27.9", optional = true }
textual = {version="^0.35.1", optional = true}
matplotlib = {version="^3.7.2", optional = true}

[tool.poetry.extras]
extras = ["networkx", "matplotlib", "openai", "textual", "streamlit"]
textual = ["textual"]
webapp = ["streamlit", "openai"]

[tool.poetry.scripts]
storytime-cli = { callable = "storytime_ai:story.simpleplay" }
storytime-checker = { callable = "storytime_ai:story.checkintegrity" }
storytime-tui = { callable = "storytime_ai:textual_app.startapp", extras = ["textual"] }
storytime = { callable = "storytime_ai:streamlit_app",  extras = ["webapp"]}

[tool.poetry.group.dev.dependencies]
pytest = "*"
pre-commit = "*"
pillow = "*"
pytest-asyncio = "*"
sphinx = "*"
sphinx-rtd-theme = "*"
myst-parser = "*"
textual = {extras = ["dev"], version = "*"}
ruff = "^0.0.286"
black = "^23.7.0"
pytest-testmon = "^2.0.12"

[tool.ruff]
line-length = 120
exclude = [ "**/__init__.py" ]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
preview = true

[tool.mypy]
ignore_missing_imports = true
files = ["**/*.py", "*.py"]
exclude = [".archiv" ]

[tool.pytest.ini_options]
minversion = "6.0"
addopts = [
    "--testmon",
    "--import-mode=importlib",
    "-ra -q",
    "--ignore=.archiv",
]
pythonpath = ["."]
testpaths = [
    "test",
]
